+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 # Unreleased
 
 - Don't run suite hooks when suite is not going to be run.
+- Gracefully shutdown even when luanit calls `os.exit`.
 
 # 0.2.0